requirements.txt
structlog~=21.5.0
unsync~=1.4.0
pygithub~=1.55